@@ -1212,7 +1212,7 @@ export class OsServer {
 
       // extract dir and save to disk in local measures directory
       // convert arraybuffer to node buffer
-      const buf = new Buffer(new Uint8Array(response.data));
+      const buf = new Buffer.from(new Uint8Array(response.data));
       if (vm.Message.showDebug()) vm.$log.debug('buffer');
 
       // save
